moved code to ui filesystem component cleanup unused methods
